Add initial data generation to app initialization
